Move `TypeckResults` to seperate module
